refactor(layout): move everything related to the layout together
