+++ src/layout/index.ts
@@ -1,0 +1,9 @@
+import useLayout from './hooks/useLayout';
+import useResize from './hooks/useResize';
+import useSplit from './hooks/useSplit';
+
+export default {
+  useLayout,
+  useSplit,
+  useResize,
+};
